add d3 format in dimension tables, additional cleanup
